- Refactored imposm3/database/postgis to imposm3/database/sql/postgis
