+++ database/sql/doc.go
@@ -1,4 +1,4 @@
 /*
 Package postgis implements the database interfaces for PostGIS.
 */
-package postgis
+package sql
