+++ diff/process.go
@@ -8,7 +8,7 @@ import (
 	"imposm3/cache"
 	"imposm3/config"
 	"imposm3/database"
-	_ "imposm3/database/postgis"
+	_ "imposm3/database/sql/postgis"
 	"imposm3/diff/parser"
 	diffstate "imposm3/diff/state"
 	"imposm3/element"
